# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ClienteEntrada": {
        "additionalProperties": false,
        "properties": {
          "apellidos": {
            "default": "",
            "maxLength": 200,
            "title": "Apellidos",
            "type": "string"
          },
          "autoriza_tratamiento": {
            "description": "autorización de tratamiento de datos (Ley 1581 de 2012)",
            "title": "Autoriza Tratamiento",
            "type": "boolean"
          },
          "digito_verificacion": {
            "anyOf": [
              {
                "pattern": "^[0-9]$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "solo NIT; si falta se calcula",
            "title": "Digito Verificacion"
          },
          "email": {
            "maxLength": 200,
            "minLength": 5,
            "title": "Email",
            "type": "string"
          },
          "nombres": {
            "description": "para NIT: razón social",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombres",
            "type": "string"
          },
          "numero_documento": {
            "maxLength": 20,
            "minLength": 3,
            "title": "Numero Documento",
            "type": "string"
          },
          "tipo_documento": {
            "enum": [
              "CC",
              "NIT",
              "NIT_EXTERIOR",
              "CE",
              "PA",
              "PPT"
            ],
            "title": "Tipo Documento",
            "type": "string"
          }
        },
        "required": [
          "tipo_documento",
          "numero_documento",
          "nombres",
          "email",
          "autoriza_tratamiento"
        ],
        "title": "ClienteEntrada",
        "type": "object"
      },
...
      "ClienteSalida": {
        "properties": {
          "activo": {
            "title": "Activo",
            "type": "boolean"
          },
          "apellidos": {
            "title": "Apellidos",
            "type": "string"
          },
          "digito_verificacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digito Verificacion"
          },
          "email_enmascarado": {
            "title": "Email Enmascarado",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre_mostrar": {
            "description": "razón social si es NIT; nombres y apellidos si es persona",
            "title": "Nombre Mostrar",
            "type": "string"
          },
          "nombres": {
            "title": "Nombres",
            "type": "string"
          },
          "numero_documento": {
            "title": "Numero Documento",
            "type": "string"
          },
          "tipo_documento": {
            "enum": [
              "CC",
              "NIT",
              "NIT_EXTERIOR",
              "CE",
              "PA",
              "PPT"
            ],
            "title": "Tipo Documento",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo_documento",
          "numero_documento",
          "digito_verificacion",
          "nombre_mostrar",
          "nombres",
          "apellidos",
          "email_enmascarado",
          "activo"
        ],
        "title": "ClienteSalida",
        "type": "object"
      },
      "ClientesPaginaSalida": {
        "properties": {
          "filas": {
            "items": {
              "$ref": "#/components/schemas/ClienteSalida"
            },
            "title": "Filas",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "filas",
          "total"
        ],
        "title": "ClientesPaginaSalida",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "detail": {
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImportacionSalida": {
        "properties": {
          "en_seco": {
            "description": "true: solo se validó; no se guardó nada",
            "title": "En Seco",
            "type": "boolean"
          },
          "existentes": {
            "description": "ya estaban registrados: no se modificaron",
            "title": "Existentes",
            "type": "integer"
          },
          "filas": {
            "title": "Filas",
            "type": "integer"
          },
          "nuevas": {
            "title": "Nuevas",
            "type": "integer"
          },
          "rechazadas": {
            "items": {
              "$ref": "#/components/schemas/FilaRechazadaSalida"
            },
            "title": "Rechazadas",
            "type": "array"
          }
        },
        "required": [
          "en_seco",
          "filas",
          "nuevas",
          "existentes",
          "rechazadas"
        ],
        "title": "ImportacionSalida",
        "type": "object"
      },
...
      "UsuarioListadoSalida": {
        "properties": {
          "activo": {
            "title": "Activo",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "rol": {
            "enum": [
              "BOMBERO",
              "ADMIN"
            ],
            "title": "Rol",
            "type": "string"
          },
          "usuario": {
            "title": "Usuario",
            "type": "string"
          }
        },
        "required": [
          "id",
          "usuario",
          "nombre",
          "rol",
          "activo"
        ],
        "title": "UsuarioListadoSalida",
        "type": "object"
      },
...
      "VehiculosPaginaSalida": {
        "properties": {
          "filas": {
            "items": {
              "$ref": "#/components/schemas/VehiculoSalida"
            },
            "title": "Filas",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "filas",
          "total"
        ],
        "title": "VehiculosPaginaSalida",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "token de POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema que reemplaza a Nexus. **Banco de pruebas: no envía nada a la DIAN.** Dinero y galones viajan como texto decimal.",
    "title": "Estación de servicio — banco de pruebas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clientes": {
      "get": {
        "operationId": "listar_clientes_clientes_get",
        "parameters": [
          {
            "description": "inicio del documento o parte del nombre",
            "in": "query",
            "name": "texto",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inicio del documento o parte del nombre",
              "title": "Texto"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "default": "activos",
              "enum": [
                "activos",
                "inactivos",
                "todos"
              ],
              "title": "Estado",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "desplazamiento",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Desplazamiento",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientesPaginaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar y buscar clientes",
        "tags": [
          "administración"
        ]
      },
      "post": {
        "operationId": "crear_clientes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registrar un cliente desde la pista (exige la autorización de datos)",
        "tags": [
          "clientes"
        ]
      }
    },
...
    "/clientes/importar": {
      "post": {
        "operationId": "importar_clientes_csv_clientes_importar_post",
        "parameters": [
          {
            "in": "query",
            "name": "en_seco",
            "required": false,
            "schema": {
              "default": true,
              "title": "En Seco",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "example": "tipo_documento;numero_documento;nombres;apellidos;email;autoriza_tratamiento;fecha_autorizacion\n",
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportacionSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Importar clientes desde CSV (en seco por defecto; exige autorización de datos con fecha)",
        "tags": [
          "administración"
        ]
      }
    },
    "/clientes/{cliente_id}/activar": {
      "post": {
        "operationId": "activar_cliente_clientes__cliente_id__activar_post",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_id",
            "required": true,
            "schema": {
              "title": "Cliente Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reactivar un cliente",
        "tags": [
          "administración"
        ]
      }
    },
    "/clientes/{cliente_id}/desactivar": {
      "post": {
        "operationId": "desactivar_cliente_clientes__cliente_id__desactivar_post",
        "parameters": [
          {
            "in": "path",
            "name": "cliente_id",
            "required": true,
            "schema": {
              "title": "Cliente Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Baja lógica: no se le factura ni aparece en la pista; se conserva su historial",
        "tags": [
          "administración"
        ]
      }
    },
...
    "/usuarios": {
      "get": {
        "operationId": "usuarios_usuarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioListadoSalida"
                  },
                  "title": "Response Usuarios Usuarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Usuarios (para filtrar por vendedor)",
        "tags": [
          "administración"
        ]
      }
    },
    "/vehiculos": {
      "get": {
        "operationId": "listar_vehiculos_vehiculos_get",
        "parameters": [
          {
            "description": "parte de la placa",
            "in": "query",
            "name": "texto",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 60,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "parte de la placa",
              "title": "Texto"
            }
          },
          {
            "in": "query",
            "name": "cliente_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cliente Id"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "default": "activos",
              "enum": [
                "activos",
                "inactivos",
                "todos"
              ],
              "title": "Estado",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "desplazamiento",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Desplazamiento",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculosPaginaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar y buscar vehículos",
        "tags": [
          "administración"
        ]
      },
      "post": {
        "operationId": "crear_vehiculo_vehiculos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registrar un vehículo o equipo de un cliente",
        "tags": [
          "administración"
        ]
      }
    },
    "/vehiculos/importar": {
      "post": {
        "operationId": "importar_vehiculos_csv_vehiculos_importar_post",
        "parameters": [
          {
            "in": "query",
            "name": "en_seco",
            "required": false,
            "schema": {
              "default": true,
              "title": "En Seco",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/csv": {
              "example": "placa;tipo_documento;numero_documento;descripcion\n",
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportacionSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Importar vehículos desde CSV (en seco por defecto; los dueños deben existir)",
        "tags": [
          "administración"
        ]
      }
    },
    "/vehiculos/{vehiculo_id}/activar": {
      "post": {
        "operationId": "activar_vehiculo_vehiculos__vehiculo_id__activar_post",
        "parameters": [
          {
            "in": "path",
            "name": "vehiculo_id",
            "required": true,
            "schema": {
              "title": "Vehiculo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Activar Vehiculo",
        "tags": [
          "administración"
        ]
      }
    },
    "/vehiculos/{vehiculo_id}/desactivar": {
      "post": {
        "operationId": "desactivar_vehiculo_vehiculos__vehiculo_id__desactivar_post",
        "parameters": [
          {
            "in": "path",
            "name": "vehiculo_id",
            "required": true,
            "schema": {
              "title": "Vehiculo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Desactivar Vehiculo",
        "tags": [
          "administración"
        ]
      }
    },
...
  }
}